+++ game/ato/flightplans/formationattack.py
@@ -178,17 +178,17 @@ class FormationAttackBuilder(IBuilder[FlightPlanT, LayoutT], ABC):
         builder = WaypointBuilder(self.flight, targets)
 
         target_waypoints: list[FlightWaypoint] = []
-        if targets is None:
+        if targets:
+            for target in targets:
+                target_waypoints.append(
+                    self.target_waypoint(self.flight, builder, target)
+                )
+        else:
             target_waypoints.append(
                 self.target_area_waypoint(
                     self.flight, self.flight.package.target, builder
                 )
             )
-        else:
-            for target in targets:
-                target_waypoints.append(
-                    self.target_waypoint(self.flight, builder, target)
-                )
 
         hold = None
         if not self.flight.is_helo: